+++ models/config/coinbase_pro_parser.py
@@ -8,7 +8,7 @@ from models.exchange.Granularity import Granularity
 
 
 def isMarketValid(market) -> bool:
-    p = re.compile(r"^[1-9A-Z]{2,20}\-[1-9A-Z]{2,20}$")
+    p = re.compile(r"^[0-9A-Z]{1,10}\-[1-9A-Z]{2,5}$")
     return p.match(market) is not None
 
 def parseMarket(market):